coding/hyperliquid_utils/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitarray==3.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
colorlog==6.10.1
cryptography==46.0.3
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperliquid-python-sdk==0.21.0
idna==3.11
msgpack==1.1.2
multidict==6.7.1
parsimonious==0.10.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.1
python-telegram-bot==22.6
regex==2026.1.15
requests==2.32.5
rlp==4.1.0
toolz==1.1.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websocket-client==1.9.0
yarl==1.22.0
gunicorn==21.2.0
uvicorn==0.30.0
psycopg2-binary==2.9.9

